+++ clang/lib/Sema/SemaDeclCXX.cpp
@@ -477,6 +477,11 @@ bool Sema::MergeCXXFunctionDecl(FunctionDecl *New, FunctionDecl *Old,
       // Ignore default arguments of old decl if they are not in
       // the same scope and this is not an out-of-line definition of
       // a member function.
+      //
+      // extern "C" functions can have default arguments across different
+      // scopes, but diagnosing that early would reject well-formed code
+      // (_N5001_.[over.best.match]/4.) Instead, they are checked
+      // in BestViableFunction after the best viable function has been selected.
       continue;
     }