+++ clang/lib/Sema/SemaOverload.cpp
@@ -10928,6 +10928,53 @@ OverloadCandidateSet::BestViableFunction(Sema &S, SourceLocation Loc,
     S.diagnoseEquivalentInternalLinkageDeclarations(Loc, Best->Function,
                                                     EquivalentCands);
 
+  // [over.match.best]/4:
+  // If the best viable function resolves to a function
+  // for which multiple declarations were found,
+  // and if any two of these declarations inhabit different scopes
+  // and specify a default argument that made the function viable,
+  // the program is ill-formed.
+  if (Best->Function && Best->Function->isExternC() &&
+      Best->ExplicitCallArguments < Best->Function->getNumParams()) {
+    // Calculate the range of parameters,
+    // default arguments of which made the candidate viable.
+    int FirstDefaultArgIndex = Best->ExplicitCallArguments;
+    int LastDefaultArgIndex = Best->Function->getNumParams() - 1;
+
+    // For each such parameter, collect all redeclarations
+    // that have non-inherited default argument.
+    llvm::SmallDenseMap<int, llvm::TinyPtrVector<ParmVarDecl *>> ParamRedecls(
+        LastDefaultArgIndex - FirstDefaultArgIndex + 1);
+    for (FunctionDecl *Redecl : Best->Function->redecls()) {
+      for (int i = FirstDefaultArgIndex; i <= LastDefaultArgIndex; ++i) {
+        ParmVarDecl *Param = Redecl->getParamDecl(i);
+        if (Param->hasDefaultArg() && !Param->hasInheritedDefaultArg())
+          ParamRedecls[i].push_back(Param);
+      }
+    }
+
+    // Emit the diagnostic if a given parameter has more than one declaration.
+    // MergeCXXFunctionDecl takes care of redeclarations of a default argument
+    // in the same scope, so if we found more than one,
+    // we assume they come from different scopes.
+    for (auto [ParamIndex, Redecls] : ParamRedecls) {
+      assert(!Redecls.empty());
+      if (Redecls.size() == 1)
+        continue;
+
+      ParmVarDecl *Param = Best->Function->getParamDecl(ParamIndex);
+      if (!Param->getDeclName().isEmpty()) {
+        S.Diag(Loc, diag::err_ovl_ambiguous_default_arg)
+            << 1 << Param->getName();
+      } else
+        S.Diag(Loc, diag::err_ovl_ambiguous_default_arg) << 0;
+      for (ParmVarDecl *Param : Redecls) {
+        S.Diag(Param->getDefaultArg()->getExprLoc(),
+               diag::note_default_argument_declared_here);
+      }
+    }
+  }
+
   return OR_Success;
 }